A bit more clean up and refactoring. Missing features for dashboard: 'my rooms' tab and main feed
